test(state): move MCP lifecycle-lock classifier properties to the source suite
The classifier properties lived in test/helpers/mcp-lifecycle-lock-properties.ts,
which imported Vitest and registered a describe block. test/mcp-lifecycle-lock.test.ts
pulled them in by side-effect import, so src/lib/state/mcp-lifecycle-lock-identity.ts
did not own the coverage of its own classifier.

Move the four unique properties into the source-local suite and delete the helper.

- 'keeps foreign-host and foreign-namespace contenders active' is not moved. Its
  host arm is already covered by 'never probes or reaps an owner from a different
  host' and its namespace arm by the PID-namespace property, both with broader
  arbitraries. Its distinct nowMs and grace values are inert: classifyMcpLifecycleLock
  reads nowMs, mtimeMs and corruptLockGraceMs only on the no-owner or wrong-sandbox
  branch (mcp-lifecycle-lock-identity.ts:209-213).
- 'keeps ownership when a cached start mismatch disappears on refresh' is replaced by
  the moved 'reaps a live PID only when a fresh identity read confirms the mismatch',
  which covers the same match case, adds the uncovered unavailable case, and asserts
  the full read sequence instead of one call.

The moved properties keep 250 runs, the recorded seed, and the 15s timeout.
No production code changes.

Refs: #8356
